chore: Extension loads json and css plugins in production mode - Fixes production webpack config - Fixes package script - State changes correctly in SQL editor - Refactoring: moving functions that transform query result data to the utils directory
