+++ src/webviews/QueryEditor/QueryPanel/QueryToolbar.tsx
@@ -61,8 +61,7 @@ const BaseActionsSection = () => {
                             menuButton={triggerProps}
                             primaryActionButton={{
                                 onClick: () => void dispatcher.runQuery(state.queryValue),
-                            }}
-                        >
+                            }}>
                             Run
                         </SplitButton>
                     )}
@@ -70,8 +69,8 @@ const BaseActionsSection = () => {
                 <MenuPopover>
                     {state.queryHistory.length === 0 && <MenuItem disabled>No history</MenuItem>}
                     {state.queryHistory.length > 0 &&
-                        state.queryHistory.map((query) => (
-                            <MenuItem onClick={() => dispatcher.insertText(query)}>
+                        state.queryHistory.map((query, index) => (
+                            <MenuItem onClick={() => dispatcher.insertText(query)} key={index}>
                                 {truncateString(query, 50)}
                             </MenuItem>
                         ))}
@@ -81,8 +80,7 @@ const BaseActionsSection = () => {
                 aria-label="Cancel"
                 icon={<StopRegular className={classes.iconStop} />}
                 disabled={!state.isExecuting}
-                onClick={() => void dispatcher.stopQuery(state.currentExecutionId)}
-            >
+                onClick={() => void dispatcher.stopQuery(state.currentExecutionId)}>
                 Cancel
             </ToolbarButton>
             <ToolbarButton aria-label="Open" icon={<FolderOpenRegular />} onClick={() => void dispatcher.openFile()}>
@@ -91,8 +89,7 @@ const BaseActionsSection = () => {
             <ToolbarButton
                 aria-label="Save query"
                 icon={<SaveRegular />}
-                onClick={() => void dispatcher.saveToFile(state.queryValue)}
-            >
+                onClick={() => void dispatcher.saveToFile(state.queryValue, 'nosql')}>
                 Save
             </ToolbarButton>
         </>
@@ -120,8 +117,11 @@ const LearnSection = () => {
                             <MenuItem>Query examples</MenuItem>
                         </MenuTrigger>
                         <MenuPopover>
-                            {samples.map((sample) => (
-                                <MenuItem disabled={state.isExecuting} onClick={() => dispatcher.insertText(sample)}>
+                            {samples.map((sample, index) => (
+                                <MenuItem
+                                    disabled={state.isExecuting}
+                                    onClick={() => dispatcher.insertText(sample)}
+                                    key={index}>
                                     {sample}
                                 </MenuItem>
                             ))}
@@ -144,8 +144,7 @@ const ConnectedActionsSection = () => {
             <ToolbarButton
                 aria-label="Disconnect"
                 icon={<DocumentMultipleRegular />}
-                onClick={() => void dispatcher.disconnectFromDatabase()}
-            >
+                onClick={() => void dispatcher.disconnectFromDatabase()}>
                 Disconnect
             </ToolbarButton>
             <ToolbarDivider />
@@ -165,8 +164,7 @@ const DisconnectedActionsSection = () => {
             aria-label="Connect"
             appearance={'primary'}
             icon={<DatabasePlugConnectedRegular className={classes.iconConnect} />}
-            onClick={() => void dispatcher.connectToDatabase()}
-        >
+            onClick={() => void dispatcher.connectToDatabase()}>
             Connect
         </ToolbarButton>
     );
