+++ webpack.config.ext.js
@@ -91,7 +91,7 @@ module.exports = (env, { mode }) => {
             new webpack.EnvironmentPlugin({
                 NODE_ENV: mode,
                 IS_BUNDLE: 'true',
-                DEVSERVER: process.env.DEVSERVER,
+                DEVSERVER: 'true',
             }),
             // Copy everything what is needed to run the extension
             // - We can't bundle everything into one file because system-dependent binaries in node_modules
